1. ft_split_one_cstm 함수 적용 2. main 파일에서 분할, minishell.h 파일 정리
